+++ packages/fe-mockserver-core/src/data/common.ts
@@ -64,6 +64,7 @@ export interface EntitySetInterface {
     getEntityInterface(entitySetName: string, tenantId: string): Promise<FileBasedMockData | undefined>;
     getMockData(tenantId: string): FileBasedMockData;
     isV4(): boolean;
+    isDraft(): boolean;
 }
 export interface DataAccessInterface {
     isV4(): boolean;
@@ -79,6 +80,7 @@ export interface DataAccessInterface {
     getMockEntitySet(
         entityTypeName: string,
         generateMockData?: boolean,
+        forceNullableValuesToNull?: boolean,
         containedEntityType?: EntityType,
         containedData?: any
     ): Promise<EntitySetInterface>;