+++ packages/fe-mockserver-core/src/mockdata/fileBasedMockData.ts
@@ -67,6 +67,9 @@ function performSimpleComparison(operator: string, mockValue: any, targetLiteral
         case 'ne':
             isValid = mockValue !== targetLiteral;
             break;
+        case 'in':
+            isValid = targetLiteral.includes(mockValue);
+            break;
         case 'eq':
         default:
             isValid = mockValue === targetLiteral;
@@ -91,18 +94,47 @@ function deleteData(currentNode: any, deepProperty: string) {
     }
     delete subNode[deepPropertyPath[deepPropertyPath.length - 1]];
 }
+
+function compareRowData(
+    row1: any,
+    row2: any,
+    hierarchyNodeProperty: string,
+    hierarchySourceProperty: string,
+    isDraft: boolean
+): boolean {
+    if (isDraft) {
+        if (hierarchyNodeProperty === hierarchySourceProperty) {
+            return (
+                getData(row1, hierarchyNodeProperty) === getData(row2, hierarchySourceProperty) &&
+                getData(row1, 'IsActiveEntity') === getData(row2, 'IsActiveEntity')
+            );
+        } else {
+            return (
+                getData(row1, hierarchyNodeProperty) === getData(row2, hierarchySourceProperty) &&
+                getData(row1, 'IsActiveEntity') === true
+            );
+        }
+    }
+    return getData(row1, hierarchyNodeProperty) === getData(row2, hierarchySourceProperty);
+}
 export class FileBasedMockData {
     protected _mockData: object[];
     protected _hierarchyTree: Record<string, Record<string, any>> = {};
     protected _entityType: EntityType;
     protected _mockDataEntitySet: EntitySetInterface;
     protected _contextId: string;
+    protected __generateMockData: boolean;
+    protected __forceNullableValuesToNull: boolean;
     constructor(mockData: any[], entityType: EntityType, mockDataEntitySet: EntitySetInterface, contextId: string) {
         this._entityType = entityType;
         this._contextId = contextId;
 
         this._mockDataEntitySet = mockDataEntitySet;
+        if ((mockData as any)?.__forceNullableValuesToNull) {
+            this.__forceNullableValuesToNull = true;
+        }
         if (mockData.length === 0 && (mockData as any).__generateMockData) {
+            this.__generateMockData = true;
             this._mockData = this.generateMockData();
         } else {
             this._mockData = cloneDeep(mockData);
@@ -118,8 +150,16 @@ export class FileBasedMockData {
 
     private validateProperties(mockEntry: any, properties: Property[]) {
         properties.forEach((prop) => {
-            if (!prop.nullable && !mockEntry.hasOwnProperty(prop.name)) {
-                mockEntry[prop.name] = this.getDefaultValueFromType(prop.type, prop.targetType, prop.defaultValue);
+            if (
+                (!this.__forceNullableValuesToNull || prop.nullable === false) &&
+                !mockEntry.hasOwnProperty(prop.name)
+            ) {
+                mockEntry[prop.name] = this.getDefaultValueFromType(
+                    prop,
+                    prop.type,
+                    prop.targetType,
+                    prop.defaultValue
+                );
             } else if (mockEntry.hasOwnProperty(prop.name) && isComplexTypeDefinition(prop.targetType)) {
                 // If the property is defined from a complex type we should validate the property of the complex type
                 this.validateProperties(mockEntry[prop.name], prop.targetType.properties);
@@ -229,10 +269,14 @@ export class FileBasedMockData {
     }
 
     protected getDefaultValueFromType(
+        property: Property,
         type: string,
         complexType: ComplexType | TypeDefinition | undefined,
         defaultValue?: any
     ): any {
+        if (this.__forceNullableValuesToNull && property.nullable) {
+            return null;
+        }
         if (complexType) {
             if (complexType._type === 'ComplexType') {
                 const outData: any = {};
@@ -292,6 +336,9 @@ export class FileBasedMockData {
         propertyName: string,
         lineIndex: number
     ): any {
+        if (this.__forceNullableValuesToNull && property.nullable) {
+            return null;
+        }
         let type = property.type;
         if (complexType) {
             const outData: any = {};
@@ -371,6 +418,7 @@ export class FileBasedMockData {
         const outObj: any = {};
         this._entityType.entityProperties.forEach((property: Property) => {
             outObj[property.name] = this.getDefaultValueFromType(
+                property,
                 property.type,
                 property.targetType,
                 property.defaultValue
@@ -573,9 +621,21 @@ export class FileBasedMockData {
             case 'Edm.DateTime':
             case 'Edm.DateTimeOffset':
                 let targetDateLiteral = literal;
-                if (literal && literal.startsWith("datetime'")) {
-                    targetDateLiteral = literal.substring(9, literal.length - 1);
+                if (literal) {
+                    if (Array.isArray(literal)) {
+                        targetDateLiteral = literal.map((literalValue) => {
+                            if (literalValue.startsWith("datetime'")) {
+                                return literalValue.substring(9, literalValue.length - 1);
+                            }
+                            return literalValue;
+                        });
+                    } else {
+                        if (literal.startsWith("datetime'")) {
+                            targetDateLiteral = literal.substring(9, literal.length - 1);
+                        }
+                    }
                 }
+
                 const testValue = new Date(targetDateLiteral).getTime();
                 const mockValueDate = new Date(mockValue).getTime();
                 isValid = performSimpleComparison(operator, mockValueDate, testValue);
@@ -584,10 +644,24 @@ export class FileBasedMockData {
             case 'Edm.Guid':
             default:
                 let targetLiteral = literal;
-                if (literal && literal.startsWith("guid'")) {
-                    targetLiteral = literal.substring(5, literal.length - 1);
-                } else if (literal && literal.startsWith("'")) {
-                    targetLiteral = literal.substring(1, literal.length - 1);
+                if (literal) {
+                    if (Array.isArray(literal)) {
+                        targetLiteral = literal.map((literalValue) => {
+                            if (literalValue.startsWith("guid'")) {
+                                return literalValue.substring(5, literalValue.length - 1);
+                            } else if (literalValue.startsWith("'")) {
+                                return literalValue.substring(1, literalValue.length - 1);
+                            } else {
+                                return literalValue;
+                            }
+                        });
+                    } else {
+                        if (literal.startsWith("guid'")) {
+                            targetLiteral = literal.substring(5, literal.length - 1);
+                        } else if (literal.startsWith("'")) {
+                            targetLiteral = literal.substring(1, literal.length - 1);
+                        }
+                    }
                 }
 
                 isValid = performSimpleComparison(operator, mockValue?.toString(), targetLiteral);
@@ -641,11 +715,13 @@ export class FileBasedMockData {
         const itemPerParents: Record<string, any[]> = {};
 
         inputSet.forEach((item: any) => {
-            const parentItemValue = getData(item, hierarchyDefinition.sourceReference) ?? '';
-            if (!itemPerParents[parentItemValue]) {
-                itemPerParents[parentItemValue] = [];
+            if (!this._mockDataEntitySet.isDraft() || item.IsActiveEntity !== false) {
+                const parentItemValue = getData(item, hierarchyDefinition.sourceReference) ?? '';
+                if (!itemPerParents[parentItemValue]) {
+                    itemPerParents[parentItemValue] = [];
+                }
+                itemPerParents[parentItemValue].push(item);
             }
-            itemPerParents[parentItemValue].push(item);
         });
         this._hierarchyTree[hierarchyQualifier] = itemPerParents;
 
@@ -745,7 +821,8 @@ export class FileBasedMockData {
         hierarchyDefinition: HierarchyDefinition,
         depth: number,
         matchedChildrenCount: number,
-        matchedProperties: any[]
+        matchedProperties: any[],
+        isDraft: boolean
     ) {
         let descendantCount = 0;
         if (currentNode && (depth < 0 || depth > 0)) {
@@ -769,11 +846,9 @@ export class FileBasedMockData {
                     setData(
                         currentNode,
                         hierarchyDefinition.matchedProperty,
-                        !!matchedProperties.find((prop) => {
-                            const mainItem = getData(prop, nodeProperty);
-                            const adjustedItem = getData(currentNode, nodeProperty);
-                            return mainItem === adjustedItem;
-                        })
+                        !!matchedProperties.find((prop) =>
+                            compareRowData(prop, currentNode, nodeProperty, nodeProperty, isDraft)
+                        )
                     );
                     if (getData(currentNode, hierarchyDefinition.matchedProperty)) {
                         matchedChildrenCount++;
@@ -803,7 +878,8 @@ export class FileBasedMockData {
                     hierarchyDefinition,
                     depth - 1,
                     matchedChildrenCount,
-                    matchedProperties
+                    matchedProperties,
+                    isDraft
                 );
             }
 
@@ -826,20 +902,22 @@ export class FileBasedMockData {
         if (aggregationAnnotation) {
             const nodeProperty = getNodeProperty(aggregationAnnotation)!;
             let adjustedData = data.map((adjustedRowData: any) => {
-                const item = this._mockData.find((dataItem: any) => {
-                    const mainItem = getData(dataItem, nodeProperty);
-                    const adjustedItem = getData(adjustedRowData, nodeProperty);
-                    return mainItem === adjustedItem;
-                });
+                const item = this._mockData.find((dataItem: any) =>
+                    compareRowData(
+                        dataItem,
+                        adjustedRowData,
+                        nodeProperty,
+                        nodeProperty,
+                        this._mockDataEntitySet.isDraft()
+                    )
+                );
                 return { ...item, ...adjustedRowData, ...{ $inResultSet: true } };
             });
             const restOfData = this._mockData
                 .filter((item: any) => {
-                    return !data.find((dataItem: any) => {
-                        const mainItem = getData(dataItem, nodeProperty);
-                        const adjustedItem = getData(item, nodeProperty);
-                        return mainItem === adjustedItem;
-                    });
+                    return !data.find((dataItem: any) =>
+                        compareRowData(dataItem, item, nodeProperty, nodeProperty, this._mockDataEntitySet.isDraft())
+                    );
                 })
                 .map((item: any) => {
                     return { ...item, ...{ $inResultSet: false } };
@@ -850,11 +928,9 @@ export class FileBasedMockData {
             const sourceReference = this.getSourceReference(aggregationAnnotation);
             // TODO Considering the input set the top level node is not necessarely the root node
             const allRootNodes = adjustedData.filter((node) => {
-                const parent = adjustedData.find((parent) => {
-                    const nodeValue = getData(parent, nodeProperty);
-                    const sourceReferenceValue = getData(node, sourceReference);
-                    return nodeValue === sourceReferenceValue;
-                });
+                const parent = adjustedData.find((parent) =>
+                    compareRowData(parent, node, nodeProperty, sourceReference, this._mockDataEntitySet.isDraft())
+                );
                 return !parent || !parent.$inResultSet;
             });
             allRootNodes.sort((a) => {
@@ -921,11 +997,9 @@ export class FileBasedMockData {
 
             let outData: object[] = [];
             outItems.forEach((item: any) => {
-                const subTreeData = data.find((dataItem: any) => {
-                    const mainItem = getData(dataItem, nodeProperty);
-                    const adjustedItem = getData(item, nodeProperty);
-                    return mainItem === adjustedItem;
-                });
+                const subTreeData = data.find((dataItem: any) =>
+                    compareRowData(dataItem, item, nodeProperty, nodeProperty, this._mockDataEntitySet.isDraft())
+                );
                 const nodePropertyValue = getData(item, nodeProperty);
                 if (subTreeData) {
                     if (
@@ -976,11 +1050,9 @@ export class FileBasedMockData {
         if (aggregationAnnotation) {
             const nodeProperty = getNodeProperty(aggregationAnnotation)!;
             const adjustedData = this._mockData.map((item: any) => {
-                const adjustedRowData = hierarchyFilter.find((dataItem: any) => {
-                    const mainItem = getData(dataItem, nodeProperty);
-                    const adjustedItem = getData(item, nodeProperty);
-                    return mainItem === adjustedItem;
-                });
+                const adjustedRowData = hierarchyFilter.find((dataItem: any) =>
+                    compareRowData(dataItem, item, nodeProperty, nodeProperty, this._mockDataEntitySet.isDraft())
+                );
                 if (adjustedRowData) {
                     return { ...item, ...adjustedRowData, ...{ $inResultSet: true } };
                 } else {
@@ -996,38 +1068,36 @@ export class FileBasedMockData {
 
             const subTrees: object[] = [];
             hierarchyFilter.forEach((item: any) => {
-                const parentNodeChildren = hierarchyNodes[getData(item, sourceReference)];
+                const parentNodeChildren = hierarchyNodes[getData(item, sourceReference) ?? ''];
                 if (parentNodeChildren) {
-                    const currentNode = parentNodeChildren.find((node: any) => {
-                        const mainItem = getData(node, nodeProperty);
-                        const adjustedItem = getData(item, nodeProperty);
-                        return mainItem === adjustedItem;
-                    });
-                    if (_parameters.keepStart) {
-                        if (hierarchyDefinition.matchedProperty) {
-                            // TODO compare with lastFilterTransformationResult
-                            setData(currentNode, hierarchyDefinition.matchedProperty, true);
+                    const currentNode = parentNodeChildren.find((node: any) =>
+                        compareRowData(node, item, nodeProperty, nodeProperty, this._mockDataEntitySet.isDraft())
+                    );
+                    if (currentNode) {
+                        if (_parameters.keepStart) {
+                            if (hierarchyDefinition.matchedProperty) {
+                                // TODO compare with lastFilterTransformationResult
+                                setData(currentNode, hierarchyDefinition.matchedProperty, true);
+                            }
+                            subTrees.push(currentNode);
                         }
-                        subTrees.push(currentNode);
+                        currentNode.$children?.forEach((child: any) => {
+                            this.flattenTree(
+                                child,
+                                subTrees,
+                                nodeProperty,
+                                hierarchyDefinition,
+                                _parameters.maximumDistance
+                            );
+                        });
                     }
-                    currentNode.$children?.forEach((child: any) => {
-                        this.flattenTree(
-                            child,
-                            subTrees,
-                            nodeProperty,
-                            hierarchyDefinition,
-                            _parameters.maximumDistance
-                        );
-                    });
                 }
             });
             const outData: object[] = [];
             inputSet.forEach((item: any) => {
-                const subTreeData: any = subTrees.find((dataItem: any) => {
-                    const mainItem = getData(dataItem, nodeProperty);
-                    const adjustedItem = getData(item, nodeProperty);
-                    return mainItem === adjustedItem;
-                });
+                const subTreeData: any = subTrees.find((dataItem: any) =>
+                    compareRowData(dataItem, item, nodeProperty, nodeProperty, this._mockDataEntitySet.isDraft())
+                );
                 if (subTreeData) {
                     if (
                         hierarchyDefinition.matchedDescendantCountProperty &&
@@ -1085,6 +1155,7 @@ export class FileBasedMockData {
             matchedProperty: getPathOrPropertyPath(hierarchyAnnotation.MatchedProperty)
         };
     }
+
     async getAncestors(
         inputSet: object[],
         lastFilterTransformationResult: object[],
@@ -1101,11 +1172,9 @@ export class FileBasedMockData {
             const nodeProperty = getNodeProperty(aggregationAnnotation)!;
             const sourceReference = this.getSourceReference(aggregationAnnotation);
             const adjustedData = this._mockData.map((item: any) => {
-                const adjustedRowData = limitedHierarchy.find((dataItem: any) => {
-                    const mainItem = getData(dataItem, nodeProperty);
-                    const adjustedItem = getData(item, nodeProperty);
-                    return mainItem === adjustedItem;
-                });
+                const adjustedRowData = limitedHierarchy.find((dataItem: any) =>
+                    compareRowData(item, dataItem, nodeProperty, nodeProperty, this._mockDataEntitySet.isDraft())
+                );
                 if (adjustedRowData) {
                     return { ...item, ...adjustedRowData, ...{ $inResultSet: true } };
                 } else {
@@ -1119,12 +1188,10 @@ export class FileBasedMockData {
             });
             const ancestors: any[] = [];
             limitedHierarchy.forEach((item: any) => {
-                const parentNodeChildren = hierarchyNodes[getData(item, sourceReference)];
-                const currentNode = parentNodeChildren.find((node: any) => {
-                    const mainItem = getData(node, nodeProperty);
-                    const adjustedItem = getData(item, nodeProperty);
-                    return mainItem === adjustedItem;
-                });
+                const parentNodeChildren = hierarchyNodes[getData(item, sourceReference) ?? ''];
+                const currentNode = parentNodeChildren.find((node: any) =>
+                    compareRowData(node, item, nodeProperty, nodeProperty, this._mockDataEntitySet.isDraft())
+                );
                 if (_parameters.keepStart) {
                     this.getAncestorsOfNode(
                         currentNode,
@@ -1133,7 +1200,8 @@ export class FileBasedMockData {
                         hierarchyDefinition,
                         _parameters.maximumDistance - 1,
                         0,
-                        lastFilterTransformationResult
+                        lastFilterTransformationResult,
+                        this._mockDataEntitySet.isDraft()
                     );
                 } else if (currentNode && currentNode.$parent) {
                     this.getAncestorsOfNode(
@@ -1143,17 +1211,16 @@ export class FileBasedMockData {
                         hierarchyDefinition,
                         _parameters.maximumDistance - 1,
                         1,
-                        lastFilterTransformationResult
+                        lastFilterTransformationResult,
+                        this._mockDataEntitySet.isDraft()
                     );
                 }
             });
             const outData: object[] = [];
             inputSet.forEach((item: any) => {
-                const subTreeData = ancestors.find((dataItem: any) => {
-                    const mainItem = getData(dataItem, nodeProperty);
-                    const adjustedItem = getData(item, nodeProperty);
-                    return mainItem === adjustedItem;
-                });
+                const subTreeData = ancestors.find((dataItem: any) =>
+                    compareRowData(dataItem, item, nodeProperty, nodeProperty, this._mockDataEntitySet.isDraft())
+                );
                 if (subTreeData) {
                     outData.push({ ...item, ...subTreeData });
                 }